services/dashboard-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
pydantic==2.7.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
structlog==24.1.0
prometheus-client==0.20.0
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-exporter-otlp==1.24.0
python-multipart==0.0.22
websockets==12.0
httpx==0.27.0
